+++ macros/spells/hex.js
@@ -6,18 +6,18 @@ if (args[0].tag === "OnUse") {
   const caster = tokenOrActor.actor ? tokenOrActor.actor : tokenOrActor;
 
   if (!caster || !targetUuid) {
-    ui.notifications.warn("Hunter's Mark: no token/target selected");
-    console.error("Hunter's Mark: no token/target selected");
+    ui.notifications.warn("Hex: no token/target selected");
+    console.error("Hex: no token/target selected");
     return;
   }
 
   const effectData = {
     changes: [
       {
-        key: "flags.midi-qol.huntersMark",
+        key: "flags.midi-qol.hex",
         mode: CONST.ACTIVE_EFFECT_MODES.OVERRIDE,
         value: targetUuid,
-        priority: 20,
+        priority: 20
       }, // who is marked
       {
         key: "flags.dnd5e.DamageBonusMacro",
@@ -26,7 +26,7 @@ if (args[0].tag === "OnUse") {
         priority: 20,
       }, // macro to apply the damage
     ],
-    origin: args[0].itemUuid,
+    origin: args[0].uuid, //flag the effect as associated to the spell being cast
     disabled: false,
     duration: args[0].item.effects[0].duration,
     icon: args[0].item.img,
@@ -35,12 +35,8 @@ if (args[0].tag === "OnUse") {
   effectData.duration.startTime = game.time.worldTime;
   await caster.createEmbeddedDocuments("ActiveEffect", [effectData]);
 } else if (args[0].tag === "DamageBonus") {
-  // only weapon attacks
-  if (!["mwak", "rwak"].includes(args[0].item.data.actionType)) return {};
   const targetUuid = args[0].hitTargets[0].uuid;
-  // only on the marked target
-  if (targetUuid !== getProperty(args[0].actor.flags, "midi-qol.huntersMark")) return {};
-  const damageType = args[0].item.data.damage.parts[0][1];
+  if (targetUuid !== getProperty(args[0].actor.flags, "midi-qol.hex")) return {};
   const diceMult = args[0].isCritical ? 2 : 1;
-  return { damageRoll: `${diceMult}d6[${damageType}]`, flavor: "Hunters Mark Damage" };
+  return { damageRoll: `${diceMult}d6[necrotic]`, flavor: "Hex Damage" };
 }